[#22] refactor: survey 도메인 패키지 분리
