+++ src/main/java/com/popple/server/domain/survey/controller/SurveyAdminController.java
@@ -1,7 +1,8 @@
-package com.popple.server.domain.survey;
+package com.popple.server.domain.survey.controller;
 
 import com.popple.server.common.dto.APIDataResponse;
 import com.popple.server.domain.entity.Survey;
+import com.popple.server.domain.survey.service.SurveyService;
 import com.popple.server.domain.survey.dto.SurveyCreateReqDto;
 import com.popple.server.domain.survey.dto.SurveyRespDto;
 import lombok.RequiredArgsConstructor;
