+++ src/main/java/com/popple/server/domain/survey/dto/SurveyRespDto.java
@@ -1,7 +1,7 @@
 package com.popple.server.domain.survey.dto;
 
 import com.popple.server.domain.entity.Survey;
-import com.popple.server.domain.survey.SurveyStatus;
+import com.popple.server.domain.survey.type.SurveyStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
