+++ src/main/java/com/popple/server/domain/survey/repository/SurveyOptionRepository.java
@@ -1,4 +1,4 @@
-package com.popple.server.domain.survey;
+package com.popple.server.domain.survey.repository;
 
 import com.popple.server.domain.entity.SurveyOption;
 import org.springframework.data.jpa.repository.JpaRepository;
