+++ src/main/java/com/popple/server/domain/survey/repository/SurveyRepository.java
@@ -1,4 +1,4 @@
-package com.popple.server.domain.survey;
+package com.popple.server.domain.survey.repository;
 
 import com.popple.server.domain.entity.Survey;
 import org.springframework.data.jpa.repository.JpaRepository;
