+++ src/main/java/com/popple/server/domain/survey/type/SurveyStatus.java
@@ -1,4 +1,4 @@
-package com.popple.server.domain.survey;
+package com.popple.server.domain.survey.type;
 
 public enum SurveyStatus {
     REVERT, IN_PROGRESS, WAIT
